Step 1: Introduce static versioning of ElasticSearch logic
