+++ mailbox/elasticsearch/src/main/java/org/apache/james/mailbox/elasticsearch/MailboxMappingFactory.java
@@ -19,274 +19,10 @@
 
 package org.apache.james.mailbox.elasticsearch;
 
-import static org.apache.james.backends.es.IndexCreationFactory.CASE_INSENSITIVE;
-import static org.apache.james.backends.es.NodeMappingFactory.ANALYZER;
-import static org.apache.james.backends.es.NodeMappingFactory.BOOLEAN;
-import static org.apache.james.backends.es.NodeMappingFactory.FIELDS;
-import static org.apache.james.backends.es.NodeMappingFactory.FORMAT;
-import static org.apache.james.backends.es.NodeMappingFactory.IGNORE_ABOVE;
-import static org.apache.james.backends.es.NodeMappingFactory.INDEX;
-import static org.apache.james.backends.es.NodeMappingFactory.LONG;
-import static org.apache.james.backends.es.NodeMappingFactory.NESTED;
-import static org.apache.james.backends.es.NodeMappingFactory.NOT_ANALYZED;
-import static org.apache.james.backends.es.NodeMappingFactory.PROPERTIES;
-import static org.apache.james.backends.es.NodeMappingFactory.RAW;
-import static org.apache.james.backends.es.NodeMappingFactory.SNOWBALL;
-import static org.apache.james.backends.es.NodeMappingFactory.STRING;
-import static org.apache.james.backends.es.NodeMappingFactory.TYPE;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.BCC;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.CC;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.DATE;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.FROM;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.HAS_ATTACHMENT;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.HTML_BODY;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.IS_ANSWERED;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.IS_DELETED;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.IS_DRAFT;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.IS_FLAGGED;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.IS_RECENT;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.IS_UNREAD;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.MAILBOX_ID;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.MEDIA_TYPE;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.MESSAGE_ID;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.MODSEQ;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.SENT_DATE;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.SIZE;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.SUBJECT;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.SUBTYPE;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.TEXT;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.TEXT_BODY;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.TO;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.UID;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.USERS;
-import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.USER_FLAGS;
-import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
-
-import java.io.IOException;
-
-import org.apache.james.backends.es.NodeMappingFactory;
-import org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.EMailer;
-import org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.Property;
 import org.elasticsearch.common.xcontent.XContentBuilder;
 
-import com.google.common.base.Throwables;
+public interface MailboxMappingFactory {
 
-public class MailboxMappingFactory {
-
-    private static final int MAXIMUM_TERM_LENGTH = 4096;
-
-    public static XContentBuilder getMappingContent() {
-        try {
-            return jsonBuilder()
-                .startObject()
-
-                    .startObject(MailboxElasticSearchConstants.MESSAGE_TYPE.getValue())
-                        .startObject(PROPERTIES)
-
-                            .startObject(MESSAGE_ID)
-                                .field(TYPE, STRING)
-                                .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-
-                            .startObject(UID)
-                                .field(TYPE, LONG)
-                            .endObject()
-
-                            .startObject(MODSEQ)
-                                .field(TYPE, LONG)
-                            .endObject()
-
-                            .startObject(SIZE)
-                                .field(TYPE, LONG)
-                            .endObject()
-
-                            .startObject(IS_ANSWERED)
-                                .field(TYPE, BOOLEAN)
-                            .endObject()
-
-                            .startObject(IS_DELETED)
-                                .field(TYPE, BOOLEAN)
-                            .endObject()
-
-                            .startObject(IS_DRAFT)
-                                .field(TYPE, BOOLEAN)
-                            .endObject()
-
-                            .startObject(IS_FLAGGED)
-                                .field(TYPE, BOOLEAN)
-                            .endObject()
-
-                            .startObject(IS_RECENT)
-                                .field(TYPE, BOOLEAN)
-                            .endObject()
-
-                            .startObject(IS_UNREAD)
-                                .field(TYPE, BOOLEAN)
-                            .endObject()
-
-                            .startObject(DATE)
-                                .field(TYPE, NodeMappingFactory.DATE)
-                                .field(FORMAT, "yyyy-MM-dd'T'HH:mm:ssZ")
-                            .endObject()
-
-                            .startObject(SENT_DATE)
-                                .field(TYPE, NodeMappingFactory.DATE)
-                                .field(FORMAT, "yyyy-MM-dd'T'HH:mm:ssZ")
-                            .endObject()
-
-                            .startObject(MEDIA_TYPE)
-                                .field(TYPE, STRING)
-                                .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-
-                            .startObject(SUBTYPE)
-                                .field(TYPE, STRING)
-                                .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-
-                            .startObject(USER_FLAGS)
-                                .field(TYPE, STRING)
-                                .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-
-                            .startObject(FROM)
-                                .field(TYPE, NESTED)
-                                .startObject(PROPERTIES)
-                                    .startObject(EMailer.NAME)
-                                        .field(TYPE, STRING)
-                                        .startObject(FIELDS)
-                                            .startObject(RAW)
-                                                .field(TYPE, STRING)
-                                                .field(ANALYZER, CASE_INSENSITIVE)
-                                            .endObject()
-                                        .endObject()
-                                    .endObject()
-                                    .startObject(EMailer.ADDRESS)
-                                        .field(TYPE, STRING)
-                                        .field(INDEX, NOT_ANALYZED)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(SUBJECT)
-                                .field(TYPE, STRING)
-                                .startObject(FIELDS)
-                                    .startObject(RAW)
-                                        .field(TYPE, STRING)
-                                        .field(ANALYZER, CASE_INSENSITIVE)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(TO)
-                                .field(TYPE, NESTED)
-                                .startObject(PROPERTIES)
-                                    .startObject(EMailer.NAME)
-                                        .field(TYPE, STRING)
-                                        .startObject(FIELDS)
-                                            .startObject(RAW)
-                                                .field(TYPE, STRING)
-                                                .field(ANALYZER, CASE_INSENSITIVE)
-                                            .endObject()
-                                        .endObject()
-                                    .endObject()
-                                    .startObject(EMailer.ADDRESS)
-                                        .field(TYPE, STRING)
-                                        .field(INDEX, NOT_ANALYZED)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(CC)
-                                .field(TYPE, NESTED)
-                                .startObject(PROPERTIES)
-                                    .startObject(EMailer.NAME)
-                                        .field(TYPE, STRING)
-                                    .endObject()
-                                    .startObject(EMailer.ADDRESS)
-                                        .field(TYPE, STRING)
-                                        .field(INDEX, NOT_ANALYZED)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(BCC)
-                                .field(TYPE, NESTED)
-                                .startObject(PROPERTIES)
-                                    .startObject(EMailer.NAME)
-                                        .field(TYPE, STRING)
-                                    .endObject()
-                                    .startObject(EMailer.ADDRESS)
-                                        .field(TYPE, STRING)
-                                        .field(INDEX, NOT_ANALYZED)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(MAILBOX_ID)
-                                .field(TYPE, STRING)
-                                .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-
-                            .startObject(USERS)
-                                .field(TYPE, STRING)
-                                .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-
-                            .startObject(PROPERTIES)
-                                .field(TYPE, NESTED)
-                                .startObject(PROPERTIES)
-                                    .startObject(Property.NAMESPACE)
-                                        .field(TYPE, STRING)
-                                        .field(INDEX, NOT_ANALYZED)
-                                    .endObject()
-                                    .startObject(Property.NAME)
-                                        .field(TYPE, STRING)
-                                        .field(INDEX, NOT_ANALYZED)
-                                    .endObject()
-                                    .startObject(Property.VALUE)
-                                        .field(TYPE, STRING)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(TEXT_BODY)
-                                .field(TYPE, STRING)
-                                .startObject(FIELDS)
-                                    .startObject(RAW)
-                                        .field(TYPE, STRING)
-                                        .field(ANALYZER, CASE_INSENSITIVE)
-                                        .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(HTML_BODY)
-                                .field(TYPE, STRING)
-                                .startObject(FIELDS)
-                                    .startObject(RAW)
-                                        .field(TYPE, STRING)
-                                        .field(ANALYZER, CASE_INSENSITIVE)
-                                        .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
-                                    .endObject()
-                                .endObject()
-                            .endObject()
-
-                            .startObject(HAS_ATTACHMENT)
-                                .field(TYPE, BOOLEAN)
-                            .endObject()
-
-                            .startObject(TEXT)
-                                .field(TYPE, STRING)
-                                .field(ANALYZER, SNOWBALL)
-                                .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
-                            .endObject()
-                        .endObject()
-                    .endObject()
-                .endObject();
-        } catch (IOException e) {
-            throw Throwables.propagate(e);
-        }
-    }
+    XContentBuilder getMappingContent();
 
 }
