+++ mailbox/elasticsearch/src/main/java/org/apache/james/mailbox/elasticsearch/json/MessageToElasticSearchJson.java
@@ -19,69 +19,20 @@
 
 package org.apache.james.mailbox.elasticsearch.json;
 
-import java.time.ZoneId;
 import java.util.List;
 
-import javax.inject.Inject;
 import javax.mail.Flags;
 
-import org.apache.james.mailbox.MailboxManager;
 import org.apache.james.mailbox.MailboxSession.User;
-import org.apache.james.mailbox.elasticsearch.IndexAttachments;
-import org.apache.james.mailbox.extractor.TextExtractor;
 import org.apache.james.mailbox.store.mail.model.MailboxMessage;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.guava.GuavaModule;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.google.common.base.Preconditions;
 
-public class MessageToElasticSearchJson {
+public interface MessageToElasticSearchJson {
 
-    private final ObjectMapper mapper;
-    private final TextExtractor textExtractor;
-    private final ZoneId zoneId;
-    private final IndexAttachments indexAttachments;
+    String convertToJson(MailboxMessage message, List<User> users) throws JsonProcessingException;
 
-    public MessageToElasticSearchJson(TextExtractor textExtractor, ZoneId zoneId, IndexAttachments indexAttachments) {
-        this.textExtractor = textExtractor;
-        this.zoneId = zoneId;
-        this.indexAttachments = indexAttachments;
-        this.mapper = new ObjectMapper();
-        this.mapper.registerModule(new GuavaModule());
-        this.mapper.registerModule(new Jdk8Module());
-    }
+    String convertToJsonWithoutAttachment(MailboxMessage message, List<User> users) throws JsonProcessingException;
 
-    @Inject
-    public MessageToElasticSearchJson(TextExtractor textExtractor, IndexAttachments indexAttachments, MailboxManager mailboxManager) {
-        this(textExtractor, ZoneId.systemDefault(), indexAttachments);
-    }
-
-    public String convertToJson(MailboxMessage message, List<User> users) throws JsonProcessingException {
-        Preconditions.checkNotNull(message);
-
-        return mapper.writeValueAsString(IndexableMessage.builder()
-                .message(message)
-                .users(users)
-                .extractor(textExtractor)
-                .zoneId(zoneId)
-                .indexAttachments(indexAttachments)
-                .build());
-    }
-
-    public String convertToJsonWithoutAttachment(MailboxMessage message, List<User> users) throws JsonProcessingException {
-        return mapper.writeValueAsString(IndexableMessage.builder()
-                .message(message)
-                .users(users)
-                .extractor(textExtractor)
-                .zoneId(zoneId)
-                .indexAttachments(IndexAttachments.NO)
-                .build());
-    }
-
-    public String getUpdatedJsonMessagePart(Flags flags, long modSeq) throws JsonProcessingException {
-        Preconditions.checkNotNull(flags);
-        return mapper.writeValueAsString(new MessageUpdateJson(flags, modSeq));
-    }
+    String getUpdatedJsonMessagePart(Flags flags, long modSeq) throws JsonProcessingException;
 }
